fix migration bug

+++ resources/migrations/20170729061622_ProjectInit.php
@@ -45,7 +45,7 @@ class ProjectInit extends AbstractMigration
 		$this->createTable(Table::PROJECT_IMAGE_MAPS, function(Schema $schema)
 		{
 			$schema->primary('id')->comment('Primary Key');
-			$schema->tinyint('project_id')->comment('Project ID');
+			$schema->integer('project_id')->comment('Project ID');
 			$schema->varchar('image')->comment('Image');
 			$schema->text('params')->comment('Params');
